// src/application/domain/value-objects/payment/payment.status.enum.ts
export enum PaymentStatus {
  PENDING = 'PENDING',
  PAID = 'PAID',
  CANCELED = 'CANCELED',
  EXPIRED = 'EXPIRED',
  FAILED = 'FAILED',
}

export enum PaymentExternalStatus {
  APPROVED = 'approved',
  CANCELLED = 'cancelled',
  REJECTED = 'rejected',
  EXPIRED = 'expired',
  PENDING = 'pending',
}